+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/options/index.ts
@@ -1,5 +1,3 @@
-import './styles.css';
-
 export { OptionsDivider } from './options-divider';
 export { OptionsItem } from './options-item';
 export { OptionsPopup } from './options-popup';
