+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/options/types.ts
@@ -1,0 +1,12 @@
+import { type ReactNode } from 'react';
+
+export interface OptionsItemProps {
+  isModified: boolean;
+  children: ReactNode;
+}
+
+export interface OptionsPopupProps {
+  ariaLabel: string;
+  modifiedCount: number;
+  children: ReactNode;
+}
